+++ paddle/phi/backends/custom/custom_device.cc
@@ -13,15 +13,15 @@
 // limitations under the License.
 
 #include "paddle/fluid/platform/device_context.h"
-#include "paddle/phi/common/data_type.h"
-#include "paddle/phi/core/enforce_custom.h"
 
 #include "paddle/phi/backends/callback_manager.h"
+#include "paddle/phi/backends/custom/enforce_custom.h"
 #include "paddle/phi/backends/device_base.h"
 #include "paddle/phi/backends/device_guard.h"
 #include "paddle/phi/backends/device_manager.h"
 #include "paddle/phi/backends/event.h"
 #include "paddle/phi/backends/stream.h"
+#include "paddle/phi/common/data_type.h"
 
 static bool operator==(const C_Device_st& d1, const C_Device_st& d2) {
   return d1.id == d2.id;
